refactor: split main.go into regextra.go + unmarshal.go
main.go grew to 643 lines covering two distinct concerns: extraction
helpers (Find*, NamedGroups, AllNamedGroups, Replace, Validate) and
the reflection-driven unmarshal stack (Unmarshal*, populateStruct,
parseFieldTag, findGroupValue, RegexUnmarshaler, setFieldValue, time
helpers). Splitting along that line:

- regextra.go (226 lines) — package doc + extraction helpers. The
  surface most users touch.
- unmarshal.go (424 lines) — Unmarshal stack + RegexUnmarshaler
  interface + reflection-heavy internals. The surface a smaller group
  reaches for, but where the bulk of the logic lives.

Test file stays as one main_test.go for now — splitting it into
regextra_test.go + unmarshal_test.go is a separate refactor and would
make this diff harder to review. main_test.go can be split in a
follow-up PR if it ever becomes a problem.

No code changes: every line moved verbatim, imports trimmed per file
to satisfy 'imported and not used'. Public API surface unchanged.

Closes re-vse
